feat(db): move Follow from feed to accounts and implement ExtendedTimeStampedModel for Post and Comment
