+++ apps/accounts/models.py
@@ -1,8 +1,10 @@
 from django.contrib.auth.models import AbstractUser
+from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.html import mark_safe
 
 from apps.accounts.managers.users import UserManager
+from apps.core.models import TimeStampedModel
 
 
 class User(AbstractUser):
@@ -58,3 +60,43 @@ class User(AbstractUser):
         return mark_safe(
             f'<img src="{img_url}" width="40" height="40" style="border-radius:50%; object-fit: cover;" />'
         )
+
+
+class Follow(TimeStampedModel):
+    follower = models.ForeignKey(
+        User,
+        related_name="following",
+        on_delete=models.CASCADE,
+        help_text="User who is following another user.",
+    )
+    followed = models.ForeignKey(
+        User,
+        related_name="followers",
+        on_delete=models.CASCADE,
+        help_text="User who is being followed.",
+    )
+
+    class Meta:
+        verbose_name = "Follow"
+        verbose_name_plural = "Follows"
+        ordering = ["-created_date"]
+        constraints = [
+            models.UniqueConstraint(
+                fields=["follower", "followed"], name="unique_follow"
+            )
+        ]
+        indexes = [
+            models.Index(fields=["follower", "-created_date"]),
+            models.Index(fields=["followed", "-created_date"]),
+        ]
+
+    def __str__(self):
+        return f"@{self.follower} follows @{self.followed}"
+
+    def clean(self):
+        if self.follower == self.followed:
+            raise ValidationError("Users cannot follow themselves.")
+
+    def save(self, *args, **kwargs):
+        self.full_clean()
+        super().save(*args, **kwargs)
