+++ apps/accounts/urls.py
@@ -11,6 +11,7 @@ urlpatterns = [
     path("register/", views.register, name="register"),
     path("profile/edit/", views.edit_profile, name="edit_profile"),
     path("profile/<str:username>/", views.profile, name="profile"),
+    path("profile/<str:username>/follow/", views.follow, name="follow"),
     path("settings/", views.account_settings, name="settings"),
     path("settings/change-email", views.change_email, name="change-email"),
     path("settings/change-password", views.change_password, name="change-password"),
