+++ apps/feed/views.py
@@ -6,9 +6,8 @@ from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 import json
 
-from apps.accounts.models import User
 from apps.core.utils import paginate_queryset
-from apps.feed.models import Post, Comment, Reaction, Bookmark, Follow
+from apps.feed.models import Post, Comment, Reaction, Bookmark
 
 
 def index(request):
@@ -154,37 +153,6 @@ def comment(request, post_id):
     if body:
         Comment.objects.create(post=post, author=request.user, body=body)
     return redirect("feed:post", post_id=post_id)
-
-
-@require_POST
-@csrf_exempt
-def connect(request, username):
-    """
-    Follow or unfollow a user.
-    """
-    if not request.user.is_authenticated:
-        return JsonResponse(
-            {"status": "401", "response": "Log in to perform this action"}
-        )
-
-    target_user = get_object_or_404(User, username=username)
-    if target_user == request.user:
-        messages.warning(request, "You cannot follow yourself.")
-        return JsonResponse(
-            {"status": "400", "response": "You cannot follow yourself."}
-        )
-
-    follow, created = Follow.objects.get_or_create(
-        follower=request.user, followed=target_user
-    )
-
-    if not created:
-        follow.delete()
-        messages.info(request, f"You have unfollowed @{username}.")
-        return JsonResponse({"status": "201", "response": "Unfollowed"})
-
-    messages.success(request, f"You are now following @{username}.")
-    return JsonResponse({"status": "201", "response": "Followed"})
 
 
 @require_POST
